+++ lib/screen/post/post_screen_state.dart
@@ -5,7 +5,7 @@ part 'post_screen_state.freezed.dart';
 @freezed
 class PostScreenState with _$PostScreenState {
   const factory PostScreenState({
-    @Default(0) int rank,
+    @Default(0) double rank,
     @Default('') String attractionName,
     @Default(false) bool isSelected,
     @Default(false) bool isSpoiler,
